+++ crates/rslint_parser/src/numbers.rs
@@ -13,10 +13,10 @@ pub enum JsNum {
 /// Parse a js number as a string into a number.  
 pub fn parse_js_num(num: String) -> Option<JsNum> {
     let (radix, mut raw) = match num.get(0..2) {
-        Some("0x") | Some("0X") => (16, num.get(2..).unwrap()),
-        Some("0b") | Some("0B") => (2, num.get(2..).unwrap()),
-        Some("0o") | Some("0O") => (8, num.get(2..).unwrap()),
-        _ => (10, num.as_str()),
+        Some("0x") | Some("0X") => (16, num.get(2..).unwrap().replace("_", "")),
+        Some("0b") | Some("0B") => (2, num.get(2..).unwrap().replace("_", "")),
+        Some("0o") | Some("0O") => (8, num.get(2..).unwrap().replace("_", "")),
+        _ => (10, num.as_str().replace("_", "")),
     };
 
     if radix == 10 && raw.starts_with('0') {
@@ -27,7 +27,7 @@ pub fn parse_js_num(num: String) -> Option<JsNum> {
     }
 
     let bigint = if raw.get(raw.len() - 1..raw.len()) == Some("n") {
-        raw = raw.split_at(raw.len() - 1).0;
+        raw = raw.split_at(raw.len() - 1).0.to_string();
         true
     } else {
         false